move merge_async_iterators to common utils
